completed DB tests and seed tests. Ready for pull reqeust. Next steps are to build server routing & then move onto client.
