+++ server/server.js
@@ -3,5 +3,5 @@ const express = require('express');
 const bodyParser = require('body-parser');
 
 var data = seeds;
-
+//nothing to see here yet lol
 // console.log(data.slice(0,2));
